add interfaces, fuction used readonly datacontext

+++ EmailWebService/Interfaces/IEmailConfigurationModel.cs
@@ -2,11 +2,11 @@
 {
     public interface IEmailConfigurationModel
     {
-        int Id { get; set; }
+        long Id { get; set; }
         string ProviderName { get; set; }
         string SMTP { get; set; }
         string Port { get; set; }
         string Login { get; set; }
         string Password { get; set; }
-    }  
+    }
 }
